requirements.txt
sqlalchemy~=2.0.27
pytz==2024.1
pyjwt==2.8.0
fastapi~=0.109.2
uvicorn~=0.27.1
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]~=1.7.4
pymysql==1.1.0
cryptography==41.0.7
pydantic[email]~=2.6.1

starlette~=0.36.3
python-magic~=0.4.27
pytest~=8.0.1